统一分数类型为double：修复小数分数显示问题
- 修改ExaminaWebApplication项目中的PassingScore字段从int改为double
- 修改backup目录中的分数相关字段类型统一为double
- 修复分数计算逻辑，支持小数分数的正确计算和显示
- 更新UI显示格式，支持小数分数的格式化显示
- 确保所有项目中分数字段类型的一致性

修改的文件：
- ExaminaWebApplication/Models/MockExam/MockExamConfiguration.cs
- ExaminaWebApplication/Models/Api/Student/StudentMockExamDto.cs
- backup/Models/ExamModule.cs
- backup/Models/Exam.cs
- backup/ViewModels/QuestionViewModel.cs
- backup/Models/ExamQuestion.cs
- backup/ViewModels/TrainingViewModel.cs
- backup/Services/ScoreEvaluationService.cs

解决问题：
- 修复小数分数（如0.5分）显示为0分的问题
- 统一所有项目中分数字段的数据类型
- 提高分数计算的精度和准确性

+++ ExaminaWebApplication/Models/Api/Student/StudentMockExamDto.cs
@@ -258,7 +258,7 @@ public class CreateMockExamRequestDto
     /// </summary>
     [Required]
     [Range(1, 1000)]
-    public int PassingScore { get; set; } = 60;
+    public double PassingScore { get; set; } = 60;
 
     /// <summary>
     /// 是否随机题目顺序
