+++ ExaminaWebApplication/Models/MockExam/MockExamConfiguration.cs
@@ -43,7 +43,7 @@ public class MockExamConfiguration
     /// 及格分数
     /// </summary>
     [Required]
-    public int PassingScore { get; set; } = 60;
+    public double PassingScore { get; set; } = 60;
 
     /// <summary>
     /// 是否随机题目顺序
@@ -132,7 +132,7 @@ public class MockExam
     /// 及格分数
     /// </summary>
     [Required]
-    public int PassingScore { get; set; }
+    public double PassingScore { get; set; }
 
     /// <summary>
     /// 是否随机题目顺序
